Implemented step 2 and passed the selected font size back to the first activity. Fixed a bug where the launcher was expecting a float but the intent contained an integer.

+++ app/src/main/java/edu/temple/activities/DisplayActivity.kt
@@ -22,9 +22,9 @@ class DisplayActivity : AppCompatActivity() {
         //checks the result code, if OK, modifies the text
     val launcher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()){
         if(it.resultCode == RESULT_OK){
-            val newSize = it.data?.getFloatExtra(TEXT_SIZE_KEY, 22f)
+            val newSize = it.data?.getIntExtra(TEXT_SIZE_KEY, 22)
             if (newSize != null) {
-                lyricsDisplayTextView.textSize = newSize
+                lyricsDisplayTextView.textSize = newSize.toFloat()
             }
         }
     }
